changed some components to reduce files
