+++ content/references.mdx
@@ -1,4 +1,4 @@
-import SectionsGrid from '~/components/SectionsGrid.tsx'
+import ReferenceGrid from '~/components/ReferenceGrid.tsx'
 
 <Title>API Reference</Title>
 
@@ -12,4 +12,4 @@ Whether you're a seasoned Solid developer or just getting started with the libra
 
 Here are some links to get you started:
 
-<SectionsGrid/>
+<ReferenceGrid/>
