+++ src/components/ComponentGrid.tsx
@@ -1,3 +1,7 @@
+import { Link } from "@solidjs/router";
+import { For } from "solid-js";
+import { SECTION_LEAF_PAGE, SECTION_PAGE } from "~/NAV_SECTIONS";
+
 import { JSXElement } from "solid-js";
 
 const GridCard = (props: { icon?: JSXElement | null; label: string }) => {
@@ -11,4 +15,18 @@ const GridCard = (props: { icon?: JSXElement | null; label: string }) => {
   );
 };
 
-export default GridCard;
+const SectionsGrid = (props: { subSections: (SECTION_LEAF_PAGE | SECTION_PAGE)[] }) => {
+  return (
+    <div class="grid grid-cols-2 gap-2 md:grid-cols-3">
+      <For each={props.subSections}>
+        {(item) => (
+          <Link href={item["link"]}>
+            <GridCard label={item["name"]} />
+          </Link>
+        )}
+      </For>
+    </div>
+  );
+};
+
+export default SectionsGrid;
