+++ src/components/Footer.tsx
@@ -1,6 +1,18 @@
-import { HelpButton } from "./HelpButton";
 import IconEdit from "~icons/heroicons-outline/pencil-alt";
 import { useLocation } from "@solidjs/router";
+import IconDiscord from "~icons/ic/baseline-discord";
+
+export const HelpButton = () => {
+  return (
+    <a
+      class="flex items-center gap-2 hover:underline"
+      href="https://discord.com/invite/solidjs"
+    >
+      <IconDiscord /> Need help? Don't hesitate to ask us on Discord!
+    </a>
+  );
+};
+
 
 export default function Footer() {
   const location = useLocation();
