--- src/old/start/router/useIsRouting.md
@@ -1,3 +1,0 @@
-<title>useIsRouting</title>
-
-##### `useIsRouting` gives you an object containing the path params of the current route
