--- src/old/start/router/useMatch.md
@@ -1,5 +1,0 @@
-<title>useMatch</title>
-
-##### `useMatch` gives you a reactive object describing the URL the user is visiting
-
-<div class="text-xl">
