--- src/old/start/server/server.md
@@ -1,18 +1,0 @@
-<title>server</title>
-
-<ssr>
-
-- [Usage](#usage)
-  - [Loading route data on the server](#example)
-  - [Forms with server functions](/api/forms/createForm#forms-with-server-functions)
-
-</ssr>
-
-- [Reference](#reference)
-
-  - [`server(serverFn): ServerFunction`](#hello-world)
-  - [`ServerFunction(...args)`](#form-controller)
-    - [`ServerFunction.action`](#form-controller-form)
-    - [`ServerFunction.url`](#form-controller-form)
-
-- [Troublehooting](#troublehooting)
