--- src/old/start/session/index.md
@@ -1,3 +1,0 @@
-<title>Sessions</title>
-
-Sessions are an important part of websites that allow the server to identify requests coming from the same person, especially when it comes to server-side form validation or when JavaScript is not on the page. Sessions are a fundamental building block of many sites that let users "log in", including social, e-commerce, business, and educational websites.
